changed charid fields from INT to BIGINT

+++ modules/TRACKER_MODULE/tracked_users.sql
@@ -1,1 +1,1 @@
-CREATE TABLE IF NOT EXISTS tracked_users_<myname> (`uid` INT NOT NULL PRIMARY KEY, `name` VARCHAR(25) NOT NULL, `added_by` VARCHAR(25) NOT NULL, `added_dt` INT NOT NULL);
+CREATE TABLE IF NOT EXISTS tracked_users_<myname> (`uid` BIGINT NOT NULL PRIMARY KEY, `name` VARCHAR(25) NOT NULL, `added_by` VARCHAR(25) NOT NULL, `added_dt` INT NOT NULL);
