+++ modules/TRACKER_MODULE/tracking.sql
@@ -1,1 +1,1 @@
-CREATE TABLE IF NOT EXISTS tracking_<myname> (`uid` INT NOT NULL, `dt` INT NOT NULL, `event` VARCHAR(6) NOT NULL);
+CREATE TABLE IF NOT EXISTS tracking_<myname> (`uid` BIGINT NOT NULL, `dt` INT NOT NULL, `event` VARCHAR(6) NOT NULL);
