+++ modules/WHOIS_MODULE/name_history.sql
@@ -1,1 +1,1 @@
-CREATE TABLE IF NOT EXISTS name_history (`charid` INT NOT NULL, `name` VARCHAR(20) NOT NULL, `dimension` INT NOT NULL, `dt` INT NOT NULL, PRIMARY KEY (charid, name, dimension));
+CREATE TABLE IF NOT EXISTS name_history (`charid` BIGINT NOT NULL, `name` VARCHAR(20) NOT NULL, `dimension` INT NOT NULL, `dt` INT NOT NULL, PRIMARY KEY (charid, name, dimension));
